Create letterCombinations.java and added it to leetcode folder
Given a string containing digits from 2-9 inclusive, return all possible letter combinations that the number could represent. Return the answer in any order.

A mapping of digits to letters (just like on the telephone buttons) is given below. Note that 1 does not map to any letters.

link - https://leetcode.com/problems/letter-combinations-of-a-phone-number/description/